apps/replay/pyproject.toml
[project]
name = "replay"
version = "0.1.0"
description = "Replay engine for shadow-mode validation against recorded mainnet data"
requires-python = ">=3.11"
dependencies = [
    "backtest",
    "comparator",
    "gridcore",
    "grid-db",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]

[tool.uv.sources]
backtest = { workspace = true }
comparator = { workspace = true }
gridcore = { workspace = true }
grid-db = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/replay"]
